# docs/swagger.yaml
basePath: /
definitions:
...
  dto.SocialMedia:
    properties:
      name:
        type: string
      social_media_url:
        example: https://subdomain.domain.dom.ge/path
        type: string
    required:
    - name
    - social_media_url
    type: object
  dto.UserLogin:
    properties:
      email:
        example: name@org.dom.ge
        type: string
      password:
        minLength: 6
        type: string
    required:
    - email
    - password
    type: object
  dto.UserRegister:
    properties:
      age:
        example: 23
        type: integer
      email:
        example: name@org.dom.ge
        type: string
      password:
        minLength: 6
        type: string
      username:
        type: string
    required:
    - age
    - email
    - password
    - username
    type: object
  dto.UserUpdate:
    properties:
      email:
        example: name@org.dom.ge
        type: string
      username:
        type: string
    required:
    - email
    - username
    type: object
...
  responses.CreateSocialMedia:
    properties:
      created_at:
        type: string
      id:
        example: 1
        type: integer
      name:
        type: string
      social_media_url:
        example: https://subdomain.domain.dom.ge/path
        type: string
      user_id:
        type: integer
    type: object
  responses.ErrorMessage:
    properties:
      error_message:
        type: string
    type: object
  responses.GetAllSocialMedias:
    properties:
      social_medias:
        items:
          $ref: '#/definitions/responses.GetSocialMedia'
        type: array
    type: object
...
  responses.GetSocialMedia:
    properties:
      created_at:
        example: "2019-11-09T21:21:46+00:00"
        type: string
      id:
        example: 1
        type: integer
      name:
        type: string
      social_media_url:
        type: string
      updated_at:
        example: "2019-11-09T21:21:46+00:00"
        type: string
      user:
        $ref: '#/definitions/responses.UserSocialMedia'
      user_id:
        type: integer
    type: object
  responses.Message:
    properties:
      message:
        type: string
    type: object
...
  responses.UpdateSocialMedia:
    properties:
      id:
        example: 1
        type: integer
      name:
        type: string
      social_media_url:
        example: https://subdomain.domain.dom.ge/path
        type: string
      updated_at:
        example: "2019-11-09T21:21:46+00:00"
        type: string
      user_id:
        example: 1
        type: integer
    type: object
...
  responses.UserLogin:
    properties:
      token:
        example: header.payload.signature
        type: string
    type: object
  responses.UserRegister:
    properties:
      age:
        example: 23
        type: integer
      email:
        example: name@org.dom.ge
        type: string
      id:
        type: integer
      username:
        type: string
    type: object
  responses.UserSocialMedia:
    properties:
      id:
        example: 1
        type: integer
      username:
        type: string
    type: object
  responses.UserUpdate:
    properties:
      age:
        example: 23
        type: integer
      email:
        example: name@org.dom.ge
        type: string
      id:
        example: 1
        type: integer
      updated_at:
        example: "2019-11-09T21:21:46+00:00"
        type: string
      username:
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: bastianarf@google.com
    name: Bastian Arfianto
  description: API server for MyGram social media in "Scalable Webservice with Golang"
    course from Hacktiv8 × Kominfo.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: MyGram API
  version: "1.0"
paths:
...
  /socialmedias:
    get:
      consumes:
      - application/json
      description: Get social medias.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/responses.GetAllSocialMedias'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/responses.ErrorMessage'
        "500":
          description: Internal Server Error
      security:
      - BearerAuth: []
      summary: Get social medias
      tags:
      - socialMedias
    post:
      consumes:
      - application/json
      description: Create a social media associated with the logged in user.
      parameters:
      - description: JSON of the social media to be made.
        in: body
        name: socialMedia
        required: true
        schema:
          $ref: '#/definitions/dto.SocialMedia'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/responses.CreateSocialMedia'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/responses.ErrorMessage'
        "500":
          description: Internal Server Error
      security:
      - BearerAuth: []
      summary: Create a social media
      tags:
      - socialMedias
  /socialmedias/{socialMediaId}:
    delete:
      consumes:
      - application/json
      description: Delete a social media associated with logged in user.
      parameters:
      - description: ID number of the social media to be deleted
        in: path
        name: socialMediaId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/responses.Message'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/responses.ErrorMessage'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/responses.ErrorMessage'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/responses.ErrorMessage'
        "500":
          description: Internal Server Error
      security:
      - BearerAuth: []
      summary: Delete a social media
      tags:
      - socialMedias
    put:
      consumes:
      - application/json
      description: Update a social media associated with logged in user.
      parameters:
      - description: ID number of the social media
        in: path
        name: socialMediaId
        required: true
        type: integer
      - description: New JSON of the social media.
        in: body
        name: socialMedia
        required: true
        schema:
          $ref: '#/definitions/dto.SocialMedia'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/responses.UpdateSocialMedia'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/responses.ErrorMessage'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/responses.ErrorMessage'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/responses.ErrorMessage'
        "500":
          description: Internal Server Error
      security:
      - BearerAuth: []
      summary: Update a social media
      tags:
      - socialMedias
  /users:
    delete:
      consumes:
      - application/json
      description: Delete logged in user identified by their bearer token.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/responses.Message'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/responses.ErrorMessage'
        "500":
          description: Internal Server Error
      security:
      - BearerAuth: []
      summary: Delete logged in user
      tags:
      - users
    put:
      consumes:
      - application/json
      description: update logged in user identified by their bearer token.
      parameters:
      - description: New email and new username of the logged in user. Leave one of
          it empty if you want it to stay the same.
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/dto.UserUpdate'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/responses.UserUpdate'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/responses.ErrorMessage'
        "500":
          description: Internal Server Error
      security:
      - BearerAuth: []
      summary: Update logged in user
      tags:
      - users
  /users/login:
    post:
      consumes:
      - application/json
      description: Login a user.
      parameters:
      - description: JSON of the user to login. Minimum password length is 6.
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/dto.UserLogin'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/responses.UserLogin'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/responses.ErrorMessage'
        "500":
          description: Internal Server Error
      summary: Login a user
      tags:
      - users
  /users/register:
    post:
      consumes:
      - application/json
      description: Register a new user.
      parameters:
      - description: JSON of the user to be made. Minimum age is 9. Minimum password
          length is 6
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/dto.UserRegister'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/responses.UserRegister'
        "209":
          description: ""
          schema:
            $ref: '#/definitions/responses.Message'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/responses.ErrorMessage'
        "500":
          description: Internal Server Error
      summary: Register a new user
      tags:
      - users
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"